修复根据主线查询主机，  整理validator逻辑
