+++ src/scene_server/validator/type.go
@@ -1,0 +1,53 @@
+/*
+ * Tencent is pleased to support the open source community by making 蓝鲸 available.
+ * Copyright (C) 2017-2018 THL A29 Limited, a Tencent company. All rights reserved.
+ * Licensed under the MIT License (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ * http://opensource.org/licenses/MIT
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package validator
+
+import (
+	"configcenter/src/common"
+	"configcenter/src/common/errors"
+	api "configcenter/src/source_controller/api/object"
+)
+
+var innerObject = []string{common.BKInnerObjIDApp, common.BKInnerObjIDSet, common.BKInnerObjIDModule, common.BKInnerObjIDProc, common.BKInnerObjIDHost, common.BKInnerObjIDPlat} //{"app", "set", "module", "process", "host", "plat"}
+
+type IntOption struct {
+	Min string `bson:"min" json:"min"`
+	Max string `bson:"max" json:"max"`
+}
+
+type EnumVal struct {
+	ID        string `bson:"id"           json:"id"`
+	Name      string `bson:"name"         json:"name"`
+	Type      string `bson:"type"         json:"type"`
+	IsDefault bool   `bson:"is_default"   json:"is_default"`
+}
+
+type ValidMap struct {
+	ownerID      string
+	objID        string
+	objCtrl      string
+	IsRequireArr []string
+	IsOnlyArr    []string
+	KeyFileds    map[string]interface{}
+	PropertyKv   map[string]string
+	ccError      errors.DefaultCCErrorIf
+	forward      *api.ForwardParam
+}
+
+// InstRst define
+type InstRst struct {
+	Result  bool        `json:"result"`
+	Code    int         `json:"code"`
+	Message interface{} `json:"message"`
+	Data    interface{} `json:"data"`
+}
